airflow/requirements.txt
pandas
geopandas
google-cloud-storage
google-cloud-bigquery
apache-airflow[google]
wget
# apache-airflow-providers-dbt-cloud==2.2.0
dbt-core
dbt-bigquery